+++ docs/examples/UserGuide/getting_started.jl
@@ -9,11 +9,23 @@
 
 # Alternatively, `using Tidier` will import TidierDB in the above manner for you, where TidierDB functions and macros will be available as `DB.@mutate()` and so on, and the TidierData equivalent would be `@mutate()`.
 
-# The associated databased packages used to set up connections are currently as follows
+# There are two ways to connect to the database.  you can use `connect` without any need to load any additional packages.
+
+# For example
+# Connecting to MySQL
+# ```julia
+# conn = connect(:mysql; host="localhost", user="root", password="password", db="mydb")
+# ```
+# versus connecting to DuckDB
+# ```julia
+# conn = connect(:duckdb)
+# ```
+
+# Alternatively, you can use the packages outlined below to establish a connection through their respective methods.
 
 # - ClickHouse: ClickHouse.jl
 # - MySQL and MariaDB: MySQL.jl
-# - MSSQL:  ODBC.jl 
+# - MSSQL: ODBC.jl 
 # - Postgres: LibPQ.jl
 # - SQLite: SQLite.jl
 # - Athena: AWS.jl