+++ src/docstrings.jl
@@ -16,9 +16,7 @@ julia> df = DataFrame(id = [string('A' + i ÷ 26, 'A' + i % 26) for i in 0:9],
                         value = repeat(1:5, 2), 
                         percent = 0.1:0.1:1.0);
 
-julia> mem = duckdb_open(":memory:");
-
-julia> db = duckdb_connect(mem);
+julia> db = connect(:duckdb);
 
 julia> copy_to(db, df, "df_mem");
 
@@ -83,9 +81,7 @@ julia> df = DataFrame(id = [string('A' + i ÷ 26, 'A' + i % 26) for i in 0:9],
                         value = repeat(1:5, 2), 
                         percent = 0.1:0.1:1.0);
 
-julia> mem = duckdb_open(":memory:");
-
-julia> db = duckdb_connect(mem);
+julia> db = connect(:duckdb);
 
 julia> copy_to(db, df, "df_mem");
 
@@ -139,9 +135,7 @@ julia> df = DataFrame(id = [string('A' + i ÷ 26, 'A' + i % 26) for i in 0:9],
                         value = repeat(1:5, 2), 
                         percent = 0.1:0.1:1.0);
 
-julia> mem = duckdb_open(":memory:");
-
-julia> db = duckdb_connect(mem);
+julia> db = connect(:duckdb);
 
 julia> copy_to(db, df, "df_mem");
 
@@ -175,9 +169,7 @@ julia> df = DataFrame(id = [string('A' + i ÷ 26, 'A' + i % 26) for i in 0:9],
                         value = repeat(1:5, 2), 
                         percent = 0.1:0.1:1.0);
 
-julia> mem = duckdb_open(":memory:");
-
-julia> db = duckdb_connect(mem);
+julia> db = connect(:duckdb);
 
 julia> copy_to(db, df, "df_mem");
 
@@ -218,9 +210,7 @@ julia> df = DataFrame(id = [string('A' + i ÷ 26, 'A' + i % 26) for i in 0:9],
                         value = repeat(1:5, 2), 
                         percent = 0.1:0.1:1.0);
 
-julia> mem = duckdb_open(":memory:");
-
-julia> db = duckdb_connect(mem);
+julia> db = connect(:duckdb);
 
 julia> copy_to(db, df, "df_mem");
 
@@ -267,9 +257,7 @@ julia> df = DataFrame(id = [string('A' + i ÷ 26, 'A' + i % 26) for i in 0:9],
                         value = repeat(1:5, 2), 
                         percent = 0.1:0.1:1.0);
 
-julia> mem = duckdb_open(":memory:");
-
-julia> db = duckdb_connect(mem);
+julia> db = connect(:duckdb);
 
 julia> copy_to(db, df, "df_mem");
 
@@ -319,9 +307,7 @@ julia> df = DataFrame(id = [string('A' + i ÷ 26, 'A' + i % 26) for i in 0:9],
                         value = repeat(1:5, 2), 
                         percent = 0.1:0.1:1.0);
 
-julia> mem = duckdb_open(":memory:");
-
-julia> db = duckdb_connect(mem);
+julia> db = connect(:duckdb);
 
 julia> copy_to(db, df, "df_mem");
 
@@ -362,9 +348,7 @@ julia> df = DataFrame(id = [string('A' + i ÷ 26, 'A' + i % 26) for i in 0:9],
                         value = repeat(1:5, 2), 
                         percent = 0.1:0.1:1.0);
 
-julia> mem = duckdb_open(":memory:");
-
-julia> db = duckdb_connect(mem);
+julia> db = connect(:duckdb);
 
 julia> copy_to(db, df, "df_mem");
 
@@ -403,9 +387,7 @@ julia> df = DataFrame(id = [string('A' + i ÷ 26, 'A' + i % 26) for i in 0:9],
                         value = repeat(1:5, 2), 
                         percent = 0.1:0.1:1.0);
 
-julia> mem = duckdb_open(":memory:");
-
-julia> db = duckdb_connect(mem);
+julia> db = connect(:duckdb);
 
 julia> copy_to(db, df, "df_mem");
 
@@ -439,9 +421,7 @@ julia> df = DataFrame(id = [string('A' + i ÷ 26, 'A' + i % 26) for i in 0:9],
                         value = repeat(1:5, 2), 
                         percent = 0.1:0.1:1.0);
 
-julia> mem = duckdb_open(":memory:");
-
-julia> db = duckdb_connect(mem);
+julia> db = connect(:duckdb);
 
 julia> copy_to(db, df, "df_mem");
 
@@ -483,9 +463,7 @@ julia> df = DataFrame(id = [string('A' + i ÷ 26, 'A' + i % 26) for i in 0:9],
                         value = repeat(1:5, 2), 
                         percent = 0.1:0.1:1.0);
 
-julia> mem = duckdb_open(":memory:");
-
-julia> db = duckdb_connect(mem);
+julia> db = connect(:duckdb);
 
 julia> copy_to(db, df, "df_mem");
 
@@ -521,9 +499,7 @@ julia> df = DataFrame(id = [string('A' + i ÷ 26, 'A' + i % 26) for i in 0:9],
                         value = repeat(1:5, 2), 
                         percent = 0.1:0.1:1.0);
 
-julia> mem = duckdb_open(":memory:");
-
-julia> db = duckdb_connect(mem);
+julia> db = connect(:duckdb);
 
 julia> copy_to(db, df, "df_mem");
 
@@ -590,16 +566,14 @@ julia> df2 = DataFrame(id2 = ["AA", "AC", "AE", "AG", "AI", "AK", "AM"],
                 category = ["X", "Y", "X", "Y", "X", "Y", "X"],
                 score = [88, 92, 77, 83, 95, 68, 74]);
 
-julia> mem = duckdb_open(":memory:");
-
-julia> db = duckdb_connect(mem);
+julia> db = connect(:duckdb);
 
 julia> copy_to(db, df, "df_mem");
 
 julia> copy_to(db, df2, "df_join");
 
 julia> @chain db_table(db, :df_mem) begin
-         @left_join(:df_join, id2, id)
+         @left_join(df_join, id2, id)
          @collect
        end
 10×7 DataFrame
@@ -645,16 +619,14 @@ julia> df2 = DataFrame(id2 = ["AA", "AC", "AE", "AG", "AI", "AK", "AM"],
                 category = ["X", "Y", "X", "Y", "X", "Y", "X"],
                 score = [88, 92, 77, 83, 95, 68, 74]);
 
-julia> mem = duckdb_open(":memory:");
-
-julia> db = duckdb_connect(mem);
+julia> db = connect(:duckdb);
 
 julia> copy_to(db, df, "df_mem");
 
 julia> copy_to(db, df2, "df_join");
 
 julia> @chain db_table(db, :df_mem) begin
-         @right_join(:df_join, id2, id)
+         @right_join(df_join, id2, id)
          @collect
        end
 7×7 DataFrame
@@ -697,16 +669,14 @@ julia> df2 = DataFrame(id2 = ["AA", "AC", "AE", "AG", "AI", "AK", "AM"],
                 category = ["X", "Y", "X", "Y", "X", "Y", "X"],
                 score = [88, 92, 77, 83, 95, 68, 74]);
 
-julia> mem = duckdb_open(":memory:");
-
-julia> db = duckdb_connect(mem);
+julia> db = connect(:duckdb);
 
 julia> copy_to(db, df, "df_mem");
 
 julia> copy_to(db, df2, "df_join");
 
 julia> @chain db_table(db, :df_mem) begin
-         @inner_join(:df_join, id2, id)
+         @inner_join(df_join, id2, id)
          @collect
        end
 5×7 DataFrame
@@ -720,6 +690,157 @@ julia> @chain db_table(db, :df_mem) begin
    5 │ AI       bb            4       0.9  AI       X             95
 ```
 """
+const docstring_full_join =
+"""
+    @inner_join(sql_query, join_table, new_table_col, orignal_table_col)
+
+Perform an full join between two SQL queries based on a specified condition. 
+This syntax here is slightly different than TidierData.jl, however, because 
+SQL does not drop the joining column, for the metadata storage, it is 
+preferrable for the names to be different 
+
+# Arguments
+- `sql_query`: The primary SQL query to operate on.
+- `join_table`: The secondary SQL table to join with the primary query table.
+- `new_table_col`: Column from the new table that matches for join. 
+- `orignal_table_col`: Column from the original table that matches for join. 
+
+# Examples
+```jldoctest
+julia> df = DataFrame(id = [string('A' + i ÷ 26, 'A' + i % 26) for i in 0:9], 
+                        groups = [i % 2 == 0 ? "aa" : "bb" for i in 1:10], 
+                        value = repeat(1:5, 2), 
+                        percent = 0.1:0.1:1.0);
+
+julia> df2 = DataFrame(id2 = ["AA", "AC", "AE", "AG", "AI", "AK", "AM"],
+                category = ["X", "Y", "X", "Y", "X", "Y", "X"],
+                score = [88, 92, 77, 83, 95, 68, 74]);
+
+julia> db = connect(:duckdb);
+
+julia> copy_to(db, df, "df_mem");
+
+julia> copy_to(db, df2, "df_join");
+
+julia> @chain db_table(db, :df_mem) begin
+         @full_join(df_join, id2, id)
+         @collect
+       end
+12×7 DataFrame
+ Row │ id       groups   value    percent    id2      category  score   
+     │ String?  String?  Int64?   Float64?   String?  String?   Int64?  
+─────┼──────────────────────────────────────────────────────────────────
+   1 │ AA       bb             1        0.1  AA       X              88
+   2 │ AC       bb             3        0.3  AC       Y              92
+   3 │ AE       bb             5        0.5  AE       X              77
+   4 │ AG       bb             2        0.7  AG       Y              83
+   5 │ AI       bb             4        0.9  AI       X              95
+   6 │ AB       aa             2        0.2  missing  missing   missing 
+   7 │ AD       aa             4        0.4  missing  missing   missing 
+   8 │ AF       aa             1        0.6  missing  missing   missing 
+   9 │ AH       aa             3        0.8  missing  missing   missing 
+  10 │ AJ       aa             5        1.0  missing  missing   missing 
+  11 │ missing  missing  missing  missing    AK       Y              68
+  12 │ missing  missing  missing  missing    AM       X              74
+```
+"""
+
+const docstring_semi_join =
+"""
+    @semi_join(sql_query, join_table, new_table_col, orignal_table_col)
+
+Perform an semi join between two SQL queries based on a specified condition. 
+This syntax here is slightly different than TidierData.jl, however, because 
+SQL does not drop the joining column, for the metadata storage, it is 
+preferrable for the names to be different 
+
+# Arguments
+- `sql_query`: The primary SQL query to operate on.
+- `join_table`: The secondary SQL table to join with the primary query table.
+- `new_table_col`: Column from the new table that matches for join. 
+- `orignal_table_col`: Column from the original table that matches for join. 
+
+# Examples
+```jldoctest
+julia> df = DataFrame(id = [string('A' + i ÷ 26, 'A' + i % 26) for i in 0:9], 
+                        groups = [i % 2 == 0 ? "aa" : "bb" for i in 1:10], 
+                        value = repeat(1:5, 2), 
+                        percent = 0.1:0.1:1.0);
+
+julia> df2 = DataFrame(id2 = ["AA", "AC", "AE", "AG", "AI", "AK", "AM"],
+                category = ["X", "Y", "X", "Y", "X", "Y", "X"],
+                score = [88, 92, 77, 83, 95, 68, 74]);
+
+julia> db = connect(:duckdb);
+
+julia> copy_to(db, df, "df_mem");
+
+julia> copy_to(db, df2, "df_join");
+
+julia> @chain db_table(db, :df_mem) begin
+         @semi_join(df_join, id2, id)
+         @collect
+       end
+5×4 DataFrame
+ Row │ id       groups   value   percent  
+     │ String?  String?  Int64?  Float64? 
+─────┼────────────────────────────────────
+   1 │ AA       bb            1       0.1
+   2 │ AC       bb            3       0.3
+   3 │ AE       bb            5       0.5
+   4 │ AG       bb            2       0.7
+   5 │ AI       bb            4       0.9
+```
+"""
+
+const docstring_anti_join =
+"""
+    @anti_join(sql_query, join_table, new_table_col, orignal_table_col)
+
+Perform an anti join between two SQL queries based on a specified condition. 
+This syntax here is slightly different than TidierData.jl, however, because 
+SQL does not drop the joining column, for the metadata storage, it is 
+preferrable for the names to be different 
+
+# Arguments
+- `sql_query`: The primary SQL query to operate on.
+- `join_table`: The secondary SQL table to join with the primary query table.
+- `new_table_col`: Column from the new table that matches for join. 
+- `orignal_table_col`: Column from the original table that matches for join. 
+
+# Examples
+```jldoctest
+julia> df = DataFrame(id = [string('A' + i ÷ 26, 'A' + i % 26) for i in 0:9], 
+                        groups = [i % 2 == 0 ? "aa" : "bb" for i in 1:10], 
+                        value = repeat(1:5, 2), 
+                        percent = 0.1:0.1:1.0);
+
+julia> df2 = DataFrame(id2 = ["AA", "AC", "AE", "AG", "AI", "AK", "AM"],
+                category = ["X", "Y", "X", "Y", "X", "Y", "X"],
+                score = [88, 92, 77, 83, 95, 68, 74]);
+
+julia> db = connect(:duckdb);
+
+julia> copy_to(db, df, "df_mem");
+
+julia> copy_to(db, df2, "df_join");
+
+julia> @chain db_table(db, :df_mem) begin
+        @anti_join(df_join, id2, id)
+        @collect
+       end
+5×4 DataFrame
+ Row │ id       groups   value   percent  
+     │ String?  String?  Int64?  Float64? 
+─────┼────────────────────────────────────
+   1 │ AB       aa            2       0.2
+   2 │ AD       aa            4       0.4
+   3 │ AF       aa            1       0.6
+   4 │ AH       aa            3       0.8
+   5 │ AJ       aa            5       1.0
+```
+"""
+
 
 const docstring_rename =
 """
@@ -738,9 +859,7 @@ julia> df = DataFrame(id = [string('A' + i ÷ 26, 'A' + i % 26) for i in 0:9],
                         value = repeat(1:5, 2), 
                         percent = 0.1:0.1:1.0);
 
-julia> mem = duckdb_open(":memory:");
-
-julia> db = duckdb_connect(mem);
+julia> db = connect(:duckdb);
 
 julia> copy_to(db, df, "df_mem");
 
@@ -781,15 +900,11 @@ julia> df = DataFrame(id = [string('A' + i ÷ 26, 'A' + i % 26) for i in 0:9],
                         value = repeat(1:5, 2), 
                         percent = 0.1:0.1:1.0);
 
-julia> mem = duckdb_open(":memory:");
-
-julia> db = duckdb_connect(mem);
+julia> db = connect(:duckdb);
 
 julia> copy_to(db, df, "test");
 ```
 """
-
-
 
 const docstring_window_order =
 """
@@ -807,9 +922,7 @@ julia> df = DataFrame(id = [string('A' + i ÷ 26, 'A' + i % 26) for i in 0:9],
                         value = repeat(1:5, 2), 
                         percent = 0.1:0.1:1.0);
 
-julia> mem = duckdb_open(":memory:");
-
-julia> db = duckdb_connect(mem);
+julia> db = connect(:duckdb);
 
 julia> copy_to(db, df, "df_mem");
 ```
@@ -833,10 +946,43 @@ julia> df = DataFrame(id = [string('A' + i ÷ 26, 'A' + i % 26) for i in 0:9],
                         value = repeat(1:5, 2), 
                         percent = 0.1:0.1:1.0);
 
-julia> mem = duckdb_open(":memory:");
-
-julia> db = duckdb_connect(mem);
+julia> db = connect(:duckdb);
 
 julia> copy_to(db, df, "df_mem");
 ```
+"""
+
+const docstring_connect = 
+"""
+    connect(backend::Symbol; kwargs...)
+
+This function establishes a database connection based on the specified backend and connection parameters and sets the SQL mode
+
+# Arguments
+- `backend`: A symbol specifying the database backend to connect to. Supported backends are:
+  - `:duckdb`, `:lite`(SQLite), `:mssql`, `mysql`(for MariaDB and MySQL), `:clickhouse`, `:postgres` 
+- `kwargs`: Keyword arguments specifying the connection parameters for the selected backend. The required parameters vary depending on the backend:
+  - MySQL:
+    - `host`: The host name or IP address of the MySQL server. Default is "localhost".
+    - `user`: The username for authentication. Default is an empty string.
+    - `password`: The password for authentication.
+    - `db`: The name of the database to connect to (optional).
+    - `port`: The port number of the MySQL server (optional).
+
+# Returns
+- A database connection object based on the selected backend.
+
+# Examples
+```julia
+# Connect to MySQL
+# conn = connect(:mysql; host="localhost", user="root", password="password", db="mydb")
+# Connect to PostgreSQL using LibPQ
+# conn = connect(:postgres; host="localhost", dbname="mydb", user="postgres", password="password")
+# Connect to ClickHouse
+# conn = connect(:clickhouse; host="localhost", port=9000, database="mydb", user="default", password="")
+# Connect to SQLite
+# conn = connect(:lite)
+# Connect to DuckDB
+julia> db = connect(:duckdb)
+DuckDB.Connection(":memory:")
 """